# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Yelements E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  user: "user",
  vendor: "vendor",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  role?: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  discount?: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateUserRequest {
  name?: string;
  address?: string;
  phone?: string;
  discount?: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  subcategories: string[];
  productCount: number;
}

export interface CreateCategoryRequest {
  name: string;
  slug: string;
  description?: string;
  imageUrl?: string;
  subcategories?: string[];
}

export interface Product {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  description: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  images?: string[];
  stock: number;
  /** @nullable */
  vendorId?: number | null;
  /** @nullable */
  vendorName?: string | null;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  isFeatured: boolean;
  /** @nullable */
  discount?: number | null;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  price: number;
  originalPrice?: number;
  description: string;
  categoryId: number;
  subcategory?: string;
  imageUrl?: string;
  images?: string[];
  stock: number;
  vendorId?: number;
  isFeatured?: boolean;
  discount?: number;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CartItem {
  productId: number;
  name: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  quantity: number;
  stock: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface AddToCartRequest {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemRequest {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  name: string;
  price: number;
  quantity: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Order {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  items: OrderItem[];
  total: number;
  status: string;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  createdAt: string;
}

export interface CreateOrderRequest {
  shippingAddress: string;
  paymentMethod: string;
}

export interface UpdateOrderStatusRequest {
  status: string;
}

export interface WishlistRequest {
  productId: number;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  userName: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface CreateReviewRequest {
  rating: number;
  comment?: string;
}

export interface AdminStats {
  totalUsers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  pendingOrders: number;
  lowStockProducts: number;
  recentOrders: Order[];
  topProducts: Product[];
}

export interface Vendor {
  id: number;
  name: string;
  email: string;
  productCount: number;
  createdAt: string;
}

export interface VendorStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
  products: Product[];
}

export type GetProductsParams = {
  /**
   * @nullable
   */
  categoryId?: number | null;
  /**
   * @nullable
   */
  subcategory?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  minPrice?: number | null;
  /**
   * @nullable
   */
  maxPrice?: number | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  vendorId?: number | null;
};